% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfuns.R
\name{Himmelblau_grad}
\alias{Himmelblau_grad}
\title{Himmelblau Gradient caluclator}
\usage{
Himmelblau_grad(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
value of the gradient at the given point
}
\description{
Himmelblau Gradient caluclator
}
