% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfuns.R
\name{Himmelblau_grad_descent}
\alias{Himmelblau_grad_descent}
\title{Gradient descent trajectory calculator}
\usage{
Himmelblau_grad_descent(x0, epsilon, U)
}
\arguments{
\item{x0}{initial point}

\item{epsilon}{tolerance level}

\item{U}{max step size}
}
\value{
an optimisation trajectory according to newton's method and given tolerance epsilon and max step size U
}
\description{
Gradient descent trajectory calculator
}
