% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfuns.R
\name{Himmelblau_Hessian}
\alias{Himmelblau_Hessian}
\title{Himmelblau Hessian calculator}
\usage{
Himmelblau_Hessian(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
value of the Hessian at the given point
}
\description{
Himmelblau Hessian calculator
}
