% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfuns.R
\name{Himmelblau_Newton}
\alias{Himmelblau_Newton}
\title{Newton's method trajectory calculator}
\usage{
Himmelblau_Newton(x0, epsilon)
}
\arguments{
\item{x0}{initial point}

\item{epsilon}{tolerance level}
}
\value{
an optimisation trajectory according to newton's method and given tolerance epsilon
}
\description{
Newton's method trajectory calculator
}
