% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfuns.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Title}
\usage{
plot_trajectory(t)
}
\arguments{
\item{t}{trajectory}
}
\value{
a plot of the optimisation trajetory on the Himmelblau function xlim = ylim = c(-5,5)
}
\description{
Title
}
